meerrge master into dongsanpham

+++ AdminApp/Controllers/LoaiController.cs
@@ -19,25 +19,63 @@ namespace AdminApp.Controllers
         }
         public async Task<IActionResult> GetAllLoai()
         {
-            var loais = JsonConvert.DeserializeObject<List<Loai>>(await (await httpClient.GetAsync("https://localhost:7050/api/Loai")).Content.ReadAsStringAsync());
-            return View(loais);
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
+            {
+                var loais = JsonConvert.DeserializeObject<List<Loai>>(await (await httpClient.GetAsync("https://localhost:7050/api/Loai")).Content.ReadAsStringAsync());
+                return View(loais);
+            }
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
+
         }
         public async Task<IActionResult> DetailLoai(Guid id)
         {
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
+            {
+                var loais = (JsonConvert.DeserializeObject<List<Loai>>(await (await httpClient.GetAsync("https://localhost:7050/api/Loai")).Content.ReadAsStringAsync())).FirstOrDefault(x => x.Id == id);
+                return View(loais);
+            }
+            else
+            {
 
-            var loais = (JsonConvert.DeserializeObject<List<Loai>>(await (await httpClient.GetAsync("https://localhost:7050/api/Loai")).Content.ReadAsStringAsync())).FirstOrDefault(x => x.Id == id);
-            return View(loais);
+                return RedirectToAction("Login", "Home");
+            }
+           
 
         }
         public ActionResult CreateLoai()
         {
-            return View();
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
+            {
+                return View();
+            }
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
+
         }
-		[HttpPost]
+        [HttpPost]
 		public async Task<IActionResult> CreateLoai(Loai loai)
 		{
             // Tiếp tục gửi yêu cầu HTTP đến endpoint API nếu không có lỗi
-            var json = JsonConvert.SerializeObject(loai);
+          
+
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
+            {
+                 var json = JsonConvert.SerializeObject(loai);
             var data = new StringContent(json, Encoding.UTF8, "application/json");
 
             var response = await httpClient.PostAsync("https://localhost:7050/api/Loai/CreateAsync", data);
@@ -49,46 +87,75 @@ namespace AdminApp.Controllers
             {
                 return BadRequest(response);
             }
+            }
+            else
+            {
+
+                return RedirectToAction("Login", "Home");
+            }
+
 
         }
 
 
         public async Task<IActionResult> DeleteLoai(Guid id)
         {
-            var response = await httpClient.DeleteAsync($"https://localhost:7050/api/Loai/{id}");
-
-            if (response.IsSuccessStatusCode)
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
-                // Xóa thành công
-                return RedirectToAction("GetAllLoai");
+                var response = await httpClient.DeleteAsync($"https://localhost:7050/api/Loai/{id}");
+
+                if (response.IsSuccessStatusCode)
+                {
+                    // Xóa thành công
+                    return RedirectToAction("GetAllLoai");
+                }
+                else
+                {
+                    // Xóa không thành công, in ra kết quả để debug
+                    var responseContent = await response.Content.ReadAsStringAsync();
+                    // Log hoặc in ra lỗi để kiểm tra
+                    Console.WriteLine($"Error deleting Loai: {responseContent}");
+
+                    // Thêm thông báo lỗi vào ViewData hoặc TempData để hiển thị trong View
+                    TempData["ErrorMessage"] = "Không thể xóa loại. Vui lòng thử lại.";
+
+                    return RedirectToAction("GetAllLoai");
+                }
             }
             else
             {
-                // Xóa không thành công, in ra kết quả để debug
-                var responseContent = await response.Content.ReadAsStringAsync();
-                // Log hoặc in ra lỗi để kiểm tra
-                Console.WriteLine($"Error deleting Loai: {responseContent}");
 
-                // Thêm thông báo lỗi vào ViewData hoặc TempData để hiển thị trong View
-                TempData["ErrorMessage"] = "Không thể xóa loại. Vui lòng thử lại.";
-
-                return RedirectToAction("GetAllLoai");
+                return RedirectToAction("Login", "Home");
             }
+
         }
 
 
         public async Task<IActionResult> EditLoai(Guid Id)
         {
-            var response = await httpClient.GetAsync($"https://localhost:7050/api/Loai/{Id}");
+            
 
-            if (response.IsSuccessStatusCode)
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
-                var jsonString = await response.Content.ReadAsStringAsync();
-                var loai = JsonConvert.DeserializeObject<Loai>(jsonString);
+                var response = await httpClient.GetAsync($"https://localhost:7050/api/Loai/{Id}");
 
-                if (loai != null)
+                if (response.IsSuccessStatusCode)
                 {
-                    return View(loai);
+                    var jsonString = await response.Content.ReadAsStringAsync();
+                    var loai = JsonConvert.DeserializeObject<Loai>(jsonString);
+
+                    if (loai != null)
+                    {
+                        return View(loai);
+                    }
+                    else
+                    {
+                        return NotFound();
+                    }
                 }
                 else
                 {
@@ -97,29 +164,45 @@ namespace AdminApp.Controllers
             }
             else
             {
-                return NotFound();
+
+                return RedirectToAction("Login", "Home");
             }
+
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> EditLoai(Loai loai)
         {
-            var apiUrl = $"https://localhost:7050/api/Loai/{loai.Id}";
+            
 
-            var jsonContent = JsonConvert.SerializeObject(loai);
-            var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
 
-            var response = await httpClient.PutAsync(apiUrl, content);
-
-            if (response.IsSuccessStatusCode)
+            var accessToken = HttpContext.Session.GetString("AccessToken");
+            var accessRole = HttpContext.Session.GetString("Result");
+            if (!string.IsNullOrEmpty(accessToken) && accessRole == "Admin" || !string.IsNullOrEmpty(accessToken) && accessRole == "NhanVien")
             {
-                return RedirectToAction("GetAllLoai");
+                var apiUrl = $"https://localhost:7050/api/Loai/{loai.Id}";
+
+                var jsonContent = JsonConvert.SerializeObject(loai);
+                var content = new StringContent(jsonContent, Encoding.UTF8, "application/json");
+
+                var response = await httpClient.PutAsync(apiUrl, content);
+
+                if (response.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("GetAllLoai");
+                }
+                else
+                {
+                    return BadRequest();
+                }
             }
             else
             {
-                return BadRequest();
+
+                return RedirectToAction("Login", "Home");
             }
+
         }
 
     }
